requirements.txt
Pillow==9.5.0
PyPDF2==3.0.1
pyspark==3.4.0
sentence_transformers==2.2.2
tqdm==4.64.1
affinda==4.3.2
altair==4.2.2
attrs==22.2.0
azure-core==1.26.3
azure-identity==1.12.0
blinker==1.5
cachetools==5.3.0
certifi
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
jinja2==3.0.3
itsdangerous==2.0.1
Flask==1.1.1
jsonschema==4.17.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
portalocker==2.7.0
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.18.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
zipp==3.13.0
pymongo==4.3.3
pinecone-client==2.2.1
sentence-transformers==2.2.2
PyPDF2==1.26.0
dnspython==2.1.0